Improves the Json response handling
Replaces Core\Listener\JsonViewModelListener with Core\Listener\EnforceJsonResponseListener
which does the same but additionally checks for a route parameter names "forceJson" which when set
to true forces the use of a \Zend\View\Model\JsonModel.
It will transfer all variables from the old model to the new JsonModel.
In case of a dispatch error event, the message variable is transferred, and if display_exception is set,
the exception message is stored in the models' exception variable.