feat: add router script for php server
When developing with the internal php server, all Urls containing a "."
are considered files by the server.
This leads to errors in YAWIK, because some url routes contain dots.
The router script checks, if a file exists and "redirects" all other
cases to the "index.php"

Usage is:
php -S localhost:8000 -t public/ public/router.php